requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
requests==2.26.0
Pillow
pytesseract==0.3.10
transformers
python-dotenv==1.0.0
markupsafe==2.1.2
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2 